pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "kredo"
dynamic = ["version"]
description = "Portable agent attestation protocol — Ed25519-signed skill certifications"
readme = "SCOPE.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Jim Motes"},
    {name = "Vanguard"},
]
dependencies = [
    "pynacl>=1.5.0",
    "pydantic>=2.0",
    "typer>=0.9.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
kredo = "kredo.cli:_cli"

[tool.setuptools.dynamic]
version = {file = "VERSION"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
kredo = ["../data/*.json"]

[tool.pytest.ini_options]
testpaths = ["tests"]
